Splitted source into separate files for each node
