+++ heos/heos-player-state.html
@@ -1,0 +1,135 @@
+<script src="resources/@jn-se/node-red-contrib-heos/heos-admin-ui.js"></script>
+<link rel="stylesheet" href="resources/@jn-se/node-red-contrib-heos/heos-admin-ui.css">
+
+<script type="text/javascript">
+
+    RED.nodes.registerType('heos-player-state',{
+        category: 'HEOS',
+        color: '#E7E7AE',
+        defaults: {
+            name: {value: "", required: false},
+            device: {value:"", required: false, type:"heos-device"},
+            deviceAutodiscover: {value: "autodiscover", required: true},
+            playerId: {value: "", required: false},
+            playerState: {value: "play", required: true},
+            playerIdFromPayload: {value: true, required: false},
+            playerStateFromPayload: {value: true, required: false}
+        },
+        inputs:1,
+        outputs:1,
+        icon: "font-awesome/fa-play",
+        label: function() {
+            return this.name||"HEOS Player State";
+        },
+        paletteLabel: "Player State",
+        inputLabels: "player and state configuration",
+        outputLabels: "executed configuration",
+        oneditprepare: function() {
+
+            $("#node-input-deviceAutodiscover").typedInput({
+                types: [
+                    {
+                        value: "deviceAutodiscover",
+                        options: [
+                            { value: "autodiscover", label: "Autodiscover and connect to HEOS"},
+                            { value: "specific", label: "Connect to specific device"},
+                        ]
+                    }
+                ]
+            })
+
+            $("#node-input-deviceAutodiscover").change(function() {
+
+                var currentValue = $("#node-input-deviceAutodiscover").val();
+
+                switch (currentValue) {
+
+                    case "autodiscover":
+                            $("#heos-device-select").hide();
+                            break;
+                    default:
+                            $("#heos-device-select").show();
+                }
+            })
+
+            $("#node-input-playerState").typedInput({
+                types: [
+                    {
+                        value: "state",
+                        options: [
+                            { value: "play", label: "Play"},
+                            { value: "pause", label: "Pause"},
+                            { value: "stop", label: "Stop"},
+                        ]
+                    }
+                ]
+            })
+        }
+    });
+
+</script>
+
+<script type="text/html" data-template-name="heos-player-state">
+    <div class="form-row">
+        <label for="node-input-name"><i class="fa fa-tag"></i> Name</label>
+        <input type="text" id="node-input-name" placeholder="Name">
+    </div>
+    <hr />
+    <div class="form-row">
+        <label for="node-input-device-deviceAutodiscover"><i class="fa fa-object-wifi"></i> Connection</label>
+        <input type="text" id="node-input-deviceAutodiscover">
+    </div>
+    <div class="form-row" id="heos-device-select">
+        <label for="node-input-device"><i class="fa fa-server"></i> Device</label>
+        <input type="text" id="node-input-device">
+    </div>
+    <hr />
+    <div class="form-row">
+        <label for="node-input-playerId"><i class="fa fa-music"></i> Player ID</label>
+        <input type="text" id="node-input-playerId">
+    </div>
+    <div class="form-row checkbox">
+        <input type="checkbox" id="node-input-playerIdFromPayload">
+        <label for="node-input-playerIdFromPayload"> Allow override from msg.payload.playerId.</label>
+    </div>
+    <hr />
+    <div class="form-row">
+        <label for="node-input-playerState"><i class="fa fa-play"></i> State</label>
+        <input type="text" id="node-input-playerState">
+    </div>
+    <div class="form-row checkbox">
+        <input type="checkbox" id="node-input-playerStateFromPayload">
+        <label for="node-input-playerStateFromPayload"> Allow override from msg.payload.playerState.</label>
+    </div>
+</script>
+
+<script type="text/html" data-help-name="heos-player-state">
+    <p>A HEOS node to control a player state.</p>
+    <h3>Inputs</h3>
+    <dl class="message-properties">
+        <dt>payload.playerId
+            <span class="property-type">string</span>
+        </dt>
+        <dd> The playerId provided via the payload overrides the playerId defined in the node configuration. </dd>
+        <dt>payload.playerState
+            <span class="property-type">string</span>
+        </dt>
+        <dd> The playerState provided via the payload overrides the playerState defined in the node configuration. The value can be one of 'play', 'pause' or 'stop'.</dd>
+    </dl>
+    <h3>Outputs</h3>
+    <ol class="node-ports">
+        <li>Standard output
+            <dl class="message-properties">
+                <dt>heos <span class="property-type">JSON</span></dt>
+                <dd>The <code>heos</code> data object of the HEOS response.</dd>
+            </dl>
+        </li>
+    </ol>
+    <h3>Details</h3>
+    <p>The <em>playerId</em> defines the HEOS player which state should be configured.
+    </p>
+    <h3>References</h3>
+    <ul>
+        <li><a target="_blank" href="https://rn.dmglobal.com/euheos/HEOS_CLI_ProtocolSpecification_2021.pdf">HEOS CLI Protocol Specification (2021)</a> - full documentation of HEOS commands.</li>
+    </ul>
+</script>
